Rename *project_controller.h -> *application_controller.h
